Do not use rpmbuild macro for qmake invocation
The provided '%qmake' macro seems to override (or somehow break) the
generated build flags. This results in some .pri-provided values getting
rewritten as empty. Callind qmake directly and letting the mkspec
templates provide the required build flags is a much better choice.

So for future reference: do not use '%qmake' for anything which builds
Qt stack, or portions thereof.