Helper script: keep static tarball names
Revert to original behaviour when choosing the tarball file names for
module sources. Drop the version string from tarball name, as this keeps
the churn with OBS slightly easier.

This is a temporary measure to keep constant updates as easy as
possible. The used version for source will be in SRPM file name in any
case, and that is only generated for a succesful build.