QtBase: make qmake depend on gdb
It seems all qmake.conf files for *nix platforms pull in
gdb_dwarf_index. This CONFIG token is the one which makes post-build
invocations to gdb and objcopy. The rules are provided by
gdb_dwarf_index.prf and thus automatically used for all builds.

The files are shipped as part of qmake package, so make that package
pull in the tools it wants to use.