QtMultimediaKit: fix qmake invocation
Qmake works blindly only if there is a single .pro file in the
directory where it is executed. More than one, and we need to specify
which project file to process.

QtMultimediaKit now has two .pro files. We could use the new (and
proper) one directly, but since the single-tree build assumes
MODULENAME.pro we'll follow the convention.