QtBase: comment out device mkspecs
The installation target for mkspecs/devices/ is not present in 5.0.0
alpha1, but apparently does come up in qtbase master. Keep the line in
the spec, but comment it out and annotate the commented line.

Remove two previously commented out package blocks. They are not needed.
Keep the xlib platform plugin block just in case.