QtWebKit: first beta1 concoction
This change brings qtwebkit5.spec up to date and should finally allow to
build it (at least for Mer).

The changes since alpha1 are numerous:

- We build without debug symbols, and try to force release build only
- We still remove tests from unpacked source tree
- We filter symbols at linktime to conserve memory
- We trick the build into using gold linker
  * This might work on x86 only, needs more testing
- We disable WebKit2 from the build
- We don't build any of the tools
- We force all components into their own shared libraries

The linker trickery is a form of voodoo. The "api" target of QtWebKit
simply will *NOT* link on 32-bit platforms if WebKit2 is enabled.
The use of non-debug objects, applying symbol filter, using gold linker
and the forced shared library builds are all intermediate steps into
getting the memory consumption down, notch by notch. Even with all of
the listed tricks, the final linking stage simply ran out of memory when
ld tried to allocate more than 32-bit userland allows.

Disabling the split-process target from the build was the last item.
With that, the build finally went through.

We also drop the experimental QtDeclarative plugin from the build. It
seemed to depend somehow on WebKit2.

The build still probably needs some weird moves to work for ARM targets.