Helper script: complete rewrite
The old helper script was starting to show its age, and modifying for
new features was starting to get error-prone.

Provide a new helper script, written from scratch with python. The
features of the new script:
- provides a useful --help output
- works better with the qt5.git structure
- tries to provide ways to deal with module-specific versions
- allow to skip changelog bump (useful for packaging iterations)
- allow to specify individiual modules properly
- saner functionality with gitrefs
  * tag/branch/commit is always specified with --gitref
- package version string can be overridden directly
- allows to skip OBS sync before package update (again, iterations)
- script defaults to qt5.git::origin/master unless otherwise specified
  by --gitref option

And overall future extensions should be somewhat easier to write.

The downside is rather more complex structure and nearly 3x size. The
price of granularity and functional sensibility is high, but should be
worth all the effort.

TODO:
- better comments
- prebuilt tarball handling (not implemented yet)
- lots of sanity checks
- defuse some of the unneeded indirections; they are artifacts from
  development and some might be better off folded directly to callers