QtWebKit: steps towards 5.0.0 release build
Include new and updated patches for QtWebKit. These do not allow to
build the module yet, but they address some of the problems we see on
the surface.

NOTE: Build will at the moment break when linking 'jsc' binary, due to
https://bugs.webkit.org/show_bug.cgi?id=100592

Skipping this item would only delay where the same bug surfaces again,
so we keep jsc as a nice canary.