Helper script: implement prebuilt tarball handling
Occasionally we may want to use prebuilt, officially released Qt5 module
tarballs instead of generating them from git. This change adds the
possibility to specify the path where official tarballs are located,
overriding the usual mode of operation.

When using prebuilt tarballs, the helper script requires to specify
which modules are to be uploaded.

To ease future testing, there is now a new option '--skip-obs-commit'.